+++ packages/sdk-react/src/components/onboarding/OnboardingEntity/OnboardingEntityIndividual/OnboardingEntityIndividual.tsx
@@ -1,11 +1,10 @@
-import { useFormContext } from 'react-hook-form';
-
 import { components } from '@/api';
 import { getIdentificationLabel } from '@/components/onboarding/helpers';
 import { RHFDatePicker } from '@/ui/RHF/RHFDatePicker';
 import { RHFTextField } from '@/ui/RHF/RHFTextField';
 import { t } from '@lingui/macro';
 import { useLingui } from '@lingui/react';
+import { useFormContext } from 'react-hook-form';
 
 type IndividualType = {
   individual: OptionalIndividualSchema;
@@ -64,6 +63,7 @@ export const OnboardingEntityIndividual = ({
           label={t(i18n)`Date of birth`}
           name="individual.date_of_birth"
           control={control}
+          valueAs="string"
         />
       )}